refactor(server.js): move forms-web-app server.js to common and use in applications-service-api too
