+++ docker-compose.yml
@@ -55,6 +55,7 @@ services:
     volumes:
       - ./package.json:/opt/app/package.json
       - ./node_modules:/opt/app/node_modules
+      - ./packages/common:/opt/app/packages/common
       - ./packages/applications-service-api:/opt/app/packages/applications-service-api
       - ./uploads:/opt/app/uploads
 #    command: npm run start:dev:debug --workspace=packages/applications-service-api
