+++ packages/applications-service-api/jest.config.js
@@ -3,7 +3,7 @@ const sharedConfig = require('../../jest.config.js');
 module.exports = {
 	...sharedConfig,
 	collectCoverageFrom: ['./src/**/*.js'],
-	coveragePathIgnorePatterns: ['node_modules', '<rootDir>/src/app.js', '<rootDir>/src/server.js'],
+	coveragePathIgnorePatterns: ['node_modules', '<rootDir>/src/app.js', '<rootDir>/src/main.js'],
 	coverageThreshold: {
 		global: {
 			branches: 60,
