+++ packages/common/src/index.js
@@ -2,10 +2,12 @@ const functional = require('./functional');
 const healthcheck = require('./health');
 const prometheus = require('./prometheus');
 const utils = require('./utils');
+const { createServer } = require('./server');
 
 module.exports = {
 	functional,
 	healthcheck,
 	prometheus,
-	utils
+	utils,
+	createServer
 };
