+++ packages/forms-web-app/src/main.js
@@ -3,7 +3,7 @@
 const { createServer } = require('@pins/common');
 
 const app = require('./app');
-const config = require('./lib/config');
+const config = require('./config');
 const logger = require('./lib/logger');
 
 createServer(app, config, logger);
